requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
Flask-Cors
Flask-JWT-Extended
passlib[bcrypt]
python-dotenv
